# api/openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Instant Chat API
  version: 0.3.0
servers:
  - url: http://127.0.0.1:8080/api/v1
    description: Local development
paths:
...
  /contacts:
    get:
      operationId: listContacts
      summary: List accepted contacts
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Accepted contacts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ContactList"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "500":
          $ref: "#/components/responses/InternalError"
  /contacts/{user_id}:
    delete:
      operationId: removeContact
      summary: Remove an accepted contact
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/UserID"
      responses:
        "204":
          description: Contact removed
        "400":
          $ref: "#/components/responses/InvalidArgument"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "404":
          $ref: "#/components/responses/NotFound"
        "500":
          $ref: "#/components/responses/InternalError"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
  parameters:
...
    UserID:
      name: user_id
      in: path
      required: true
      schema:
        $ref: "#/components/schemas/ResourceID"
  responses:
    InvalidArgument:
      description: The request body or fields are invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Unauthorized:
      description: Credentials or a session token are invalid
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    NotFound:
      description: The requested resource was not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
    InternalError:
      description: The request could not be completed
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  schemas:
...
    ResourceID:
      type: string
      pattern: "^[1-9][0-9]*$"
    Username:
      type: string
      minLength: 3
      maxLength: 32
      pattern: "^[a-z][a-z0-9_]{2,31}$"
    PublicUser:
      type: object
      additionalProperties: false
      required:
        - id
        - username
        - display_name
        - created_at
      properties:
        id:
          $ref: "#/components/schemas/ResourceID"
        username:
          $ref: "#/components/schemas/Username"
        display_name:
          type: string
        created_at:
          type: string
          format: date-time
...
    Contact:
      type: object
      additionalProperties: false
      required:
        - relationship_id
        - user
        - connected_at
      properties:
        relationship_id:
          $ref: "#/components/schemas/ResourceID"
        user:
          $ref: "#/components/schemas/PublicUser"
        connected_at:
          type: string
          format: date-time
    ContactList:
      type: object
      additionalProperties: false
      required:
        - contacts
      properties:
        contacts:
          type: array
          maxItems: 500
          items:
            $ref: "#/components/schemas/Contact"
...
    ErrorResponse:
      type: object
      additionalProperties: false
      required:
        - error
      properties:
        error:
          type: object
          additionalProperties: false
          required:
            - code
            - message
            - request_id
          properties:
            code:
              type: string
            message:
              type: string
            request_id:
              type: string